pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2six"
version = "0.1.0"
description = "Render images to terminal text using Unicode sixth-block characters (2x3)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
